feat: served existing asset is resized image already exists

+++ src/routes/api/images.ts
@@ -2,6 +2,7 @@ import express from 'express';
 import path from 'path';
 import { resizeImage } from '../../utils/imageTransforms';
 import { promises as fsPromises } from 'fs';
+import fs from 'fs';
 
 const images = express.Router();
 
@@ -10,11 +11,15 @@ images.get('/', async (req, res) => {
     const filename = req.query.filename as unknown as string;
     const height = parseInt(req.query.height as unknown as string);
     const width = parseInt(req.query.width as unknown as string);
-    const resizedImage = await resizeImage(filename, height, width);
-    await fsPromises.writeFile(
-      `public/images/resized/${filename}${height}x${width}.jpg`,
-      resizedImage
-    );
+    if (
+      !fs.existsSync(`public/images/resized/${filename}${height}x${width}.jpg`)
+    ) {
+      const resizedImage = await resizeImage(filename, height, width);
+      await fsPromises.writeFile(
+        `public/images/resized/${filename}${height}x${width}.jpg`,
+        resizedImage
+      );
+    }
     res.sendFile(
       path.resolve(`public/images/resized/${filename}${height}x${width}.jpg`)
     );
